requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==3.6.4.0
celery==5.2.7
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.2
django-celery-results==2.5.1
djangorestframework==3.14.0
idna==3.10
kombu==5.4.2
prompt_toolkit==3.0.50
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.1
redis==4.5.1
requests==2.31.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
